mangaging tests
some tests will need a special config file that travis will have to put into place.
